requirements.txt
pywebview
cryptography
openai
pynput
pyperclip
pyinstaller
python-dotenv